% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.cv.rsv}
\alias{coef.cv.rsv}
\title{Extract coefficients from cv.rsv objects}
\usage{
\method{coef}{cv.rsv}(object, ...)
}
\arguments{
\item{object}{An object of class "cv.rsv"}

\item{...}{Additional arguments (unused)}
}
\value{
Named numeric vector of average coefficient
}
\description{
Extract coefficients from cv.rsv objects
}
