% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.rsv}
\alias{coef.rsv}
\title{Extract coefficients from rsv objects}
\usage{
\method{coef}{rsv}(object, ...)
}
\arguments{
\item{object}{An object of class "rsv"}

\item{...}{Additional arguments (unused)}
}
\value{
Named numeric vector of coefficients
}
\description{
Extract coefficients from rsv objects
}
