% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{confint.rsv}
\alias{confint.rsv}
\title{Confidence intervals for rsv objects}
\usage{
\method{confint}{rsv}(object, parm = "D", level = 0.95, ...)
}
\arguments{
\item{object}{An object of class "rsv"}

\item{parm}{Parameter names (default "D")}

\item{level}{Confidence level (default 0.95)}

\item{...}{Additional arguments (unused)}
}
\value{
Matrix with confidence intervals
}
\description{
Confidence intervals for rsv objects
}
