% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{.predict_model}
\alias{.predict_model}
\title{Predict from fitted model (binary outcome - internal)}
\usage{
.predict_model(model, XR_new, spec)
}
\arguments{
\item{model}{Fitted model (ranger or fastglm)}

\item{XR_new}{New covariates}

\item{spec}{Model specification (for determining model type)}
}
\value{
Predicted probabilities
}
\description{
Internal function for predicting from different model types
}
\keyword{internal}
