% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{.predict_model_multicat}
\alias{.predict_model_multicat}
\title{Predict from fitted model (multi-category outcome - internal)}
\usage{
.predict_model_multicat(model, XR_new, y_levels, spec)
}
\arguments{
\item{model}{Fitted model (ranger or multinom)}

\item{XR_new}{New covariates}

\item{y_levels}{Levels of y}

\item{spec}{Model specification (for determining model type)}
}
\value{
Data frame with predicted probabilities (pred_Y0, pred_Y1, ...)
}
\description{
Internal function for predicting from different model types for multi-category outcomes
}
\keyword{internal}
