% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_all_models}
\alias{fit_all_models}
\title{Fit all four prediction models}
\usage{
fit_all_models(
  R_train,
  X_train = NULL,
  Y,
  D,
  S_e,
  S_o,
  R_test,
  X_test = NULL,
  X_cols = NULL,
  y_levels,
  yK = NULL,
  models,
  seed = NULL,
  num.threads = 1
)
}
\arguments{
\item{R_train}{Training remote sensing variables}

\item{X_train}{Training discrete covariates (optional)}

\item{Y}{Outcome vector (can be multi-category)}

\item{D}{Treatment vector (binary)}

\item{S_e}{Experimental sample indicator (binary)}

\item{S_o}{Observational sample indicator (binary)}

\item{R_test}{Test remote sensing variables for predictions}

\item{X_test}{Test discrete covariates (optional)}

\item{X_cols}{Column names for X covariates}

\item{yK}{Reference category for Y}

\item{models}{List of model specifications for Y, D, S_e, S_o}

\item{seed}{Random seed}

\item{num.threads}{Number of threads for parallel processing}
}
\value{
List with components:
\item{predictions}{Data frame with predicted Y (all levels), D, S_e, S_o}
\item{theta_init}{Initial treatment effect estimate}
}
\description{
Fits models for P(Y|X,R,S_o=1), P(D|X,R,S_e=1), P(S_e|X,R), P(S_o|X,R)
}
\keyword{internal}
