% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_logistic}
\alias{fit_logistic}
\title{Fit logistic regression model (binary outcome)}
\usage{
fit_logistic(XR_train, y, XR_test)
}
\arguments{
\item{XR_train}{Training covariates}

\item{y}{Response variable (binary)}

\item{XR_test}{Test covariates}
}
\value{
List with predictions and model
}
\description{
Fit logistic regression model (binary outcome)
}
\keyword{internal}
