% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_logistic_multicat}
\alias{fit_logistic_multicat}
\title{Fit multinomial logistic regression (multi-category outcome)}
\usage{
fit_logistic_multicat(XR_train, y, XR_test, y_levels)
}
\arguments{
\item{XR_train}{Training covariates}

\item{y}{Response variable (multi-category)}

\item{XR_test}{Test covariates}

\item{y_levels}{Levels of y}
}
\value{
List with predictions (data.frame with pred_Y0, pred_Y1, ...) and model
}
\description{
Fit multinomial logistic regression (multi-category outcome)
}
\keyword{internal}
