% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_one_model}
\alias{fit_one_model}
\title{Fit a single prediction model (binary outcome)}
\usage{
fit_one_model(XR_train, y, XR_test, spec, seed = NULL, num.threads = 1)
}
\arguments{
\item{XR_train}{Training covariates (X and R combined)}

\item{y}{Response variable (binary)}

\item{XR_test}{Test covariates}

\item{spec}{Model specification (list with model type and parameters)}

\item{seed}{Random seed}

\item{num.threads}{Number of threads}
}
\value{
List with predictions and fitted model
}
\description{
Fit a single prediction model (binary outcome)
}
\keyword{internal}
