% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_one_model_multicat}
\alias{fit_one_model_multicat}
\title{Fit a single prediction model for multi-category outcome}
\usage{
fit_one_model_multicat(
  XR_train,
  y,
  XR_test,
  y_levels,
  spec,
  seed = NULL,
  num.threads = 1
)
}
\arguments{
\item{XR_train}{Training covariates (X and R combined)}

\item{y}{Response variable (multi-category)}

\item{XR_test}{Test covariates}

\item{y_levels}{Levels of y}

\item{spec}{Model specification (list with model type and parameters)}

\item{seed}{Random seed}

\item{num.threads}{Number of threads}
}
\value{
List with predictions (data.frame with pred_Y0, pred_Y1, ...) and fitted model
}
\description{
Fit a single prediction model for multi-category outcome
}
\keyword{internal}
