% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_randomforest}
\alias{fit_randomforest}
\title{Fit random forest model (binary outcome)}
\usage{
fit_randomforest(XR_train, y, XR_test, spec, seed = NULL, num.threads = 1)
}
\arguments{
\item{XR_train}{Training covariates}

\item{y}{Response variable (binary)}

\item{XR_test}{Test covariates}

\item{spec}{Model specification}

\item{seed}{Random seed}

\item{num.threads}{Number of threads}
}
\value{
List with predictions and model
}
\description{
Fit random forest model (binary outcome)
}
\keyword{internal}
