% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_randomforest_multicat}
\alias{fit_randomforest_multicat}
\title{Fit random forest model (multi-category outcome)}
\usage{
fit_randomforest_multicat(
  XR_train,
  y,
  XR_test,
  y_levels,
  spec,
  seed = NULL,
  num.threads = 1
)
}
\arguments{
\item{XR_train}{Training covariates}

\item{y}{Response variable (multi-category)}

\item{XR_test}{Test covariates}

\item{y_levels}{Levels of y}

\item{spec}{Model specification}

\item{seed}{Random seed}

\item{num.threads}{Number of threads}
}
\value{
List with predictions (data.frame with pred_Y0, pred_Y1, ...) and model
}
\description{
Fit random forest model (multi-category outcome)
}
\keyword{internal}
