% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_xgboost}
\alias{fit_xgboost}
\title{Fit XGBoost model (binary outcome)}
\usage{
fit_xgboost(XR_train, y, XR_test, spec, seed = NULL)
}
\arguments{
\item{XR_train}{Training covariates}

\item{y}{Response variable (binary 0/1)}

\item{XR_test}{Test covariates}

\item{spec}{Model specification (list with xgb parameters)}

\item{seed}{Random seed}
}
\value{
List with predictions and model
}
\description{
Fit XGBoost model (binary outcome)
}
\keyword{internal}
