% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_xgboost_multicat}
\alias{fit_xgboost_multicat}
\title{Fit XGBoost model (multi-category outcome)}
\usage{
fit_xgboost_multicat(XR_train, y, XR_test, y_levels, spec, seed = NULL)
}
\arguments{
\item{XR_train}{Training covariates}

\item{y}{Response variable (multi-category)}

\item{XR_test}{Test covariates}

\item{y_levels}{Levels of y}

\item{spec}{Model specification}

\item{seed}{Random seed}
}
\value{
List with predictions (data.frame with pred_Y0, pred_Y1, ...) and model
}
\description{
Fit XGBoost model (multi-category outcome)
}
\keyword{internal}
