% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsv_helpers.R
\name{get_Delta}
\alias{get_Delta}
\title{Compute treatment and outcome variations for RSV estimation (Deltas)}
\usage{
get_Delta(observations, predictions)
}
\arguments{
\item{observations}{a data frame containing:
\describe{
  \item{Y}{Outcome variable (binary, NA where not observed)}
  \item{D}{Treatment indicator (binary, NA where not observed)}
  \item{S_e}{Experimental sample indicator (0 or 1)}
  \item{S_o}{Observational sample indicator (0 or 1)}
}}

\item{predictions}{a data frame containing:
\describe{
  \item{Y}{Predicted \eqn{P(Y = 1 \mid R, S_o = 1)}, \eqn{\text{PRED}_Y(R)}}
  \item{D}{Predicted \eqn{P(D = 1 \mid R, S_e = 1)}, \eqn{\text{PRED}_D(R)}}
  \item{S_e}{Predicted \eqn{P(S_e = 1 \mid R)}, \eqn{\text{PRED}_{S_e}(R)}}
  \item{S_o}{Predicted \eqn{P(S_o = 1 \mid R)}, \eqn{\text{PRED}_{S_o}(R)}}
}}
}
\value{
A list containing:
\describe{
  \item{e}{Treatment variation in experimental sample}
  \item{o}{Outcome variation in observational sample}
}
}
\description{
Compute treatment and outcome variations for RSV estimation (Deltas)
}
\keyword{internal}
