% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsv_helpers.R
\name{get_Delta_obs}
\alias{get_Delta_obs}
\title{Compute treatment and outcome variations for RSV estimation (Delta) from marginal probabilities only}
\usage{
get_Delta_obs(observations)
}
\arguments{
\item{observations}{a data frame containing:
\describe{
  \item{Y}{Outcome variable (binary, NA where not observed)}
  \item{D}{Treatment indicator (binary, NA where not observed)}
  \item{S_e}{Experimental sample indicator (0 or 1)}
  \item{S_o}{Observational sample indicator (0 or 1)}
}}
}
\value{
A list containing:
\describe{
  \item{e}{Treatment variation in experimental sample}
  \item{o}{Outcome variation in observational sample}
}
}
\description{
Compute treatment and outcome variations for RSV estimation (Delta) from marginal probabilities only
}
\keyword{internal}
