% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{make_folds}
\alias{make_folds}
\title{Create K-fold cross-validation assignments}
\usage{
make_folds(n, nfolds = 5, seed = NULL, clusters = NULL)
}
\arguments{
\item{n}{Sample size}

\item{nfolds}{Number of folds (default 5)}

\item{seed}{Random seed (optional)}
}
\value{
Integer vector of fold assignments (1, 2, ..., nfolds)
}
\description{
Create K-fold cross-validation assignments
}
\keyword{internal}
