% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{make_split}
\alias{make_split}
\title{Create simple train/test split}
\usage{
make_split(n, train.proportion = 0.5, seed = NULL, clusters = NULL)
}
\arguments{
\item{n}{Sample size}

\item{train.proportion}{Proportion of data for training (default 0.5)}

\item{seed}{Random seed (optional)}
}
\value{
List with components:
\item{train}{Indices for training set}
\item{test}{Indices for test set}
}
\description{
Create simple train/test split
}
\keyword{internal}
