% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsv_estimator.R
\name{rsv_from_predictions}
\alias{rsv_from_predictions}
\title{Estimate from predictions without fitting}
\usage{
rsv_from_predictions(
  Y,
  D,
  S_e,
  S_o,
  pred_Y,
  pred_D,
  pred_S_e,
  pred_S_o,
  theta_init = NULL,
  eps = 0.01,
  ml_params = list(),
  se = TRUE,
  se_params = list()
)
}
\arguments{
\item{Y}{Outcome variable (binary, \code{NA} where not observed).}

\item{D}{Treatment indicator (binary, \code{NA} where not observed).}

\item{S_e}{Experimental sample indicator (0 or 1).}

\item{S_o}{Observational sample indicator (0 or 1).}

\item{pred_Y}{(Optional) Predicted \eqn{P[Y \mid R, S_o = 1]}, \eqn{\text{PRED}_Y(R)}. If provided, other predictions must also be provided.}

\item{pred_D}{(Optional) Predicted \eqn{P[D \mid R, S_e = 1]}, \eqn{\text{PRED}_D(R)}.}

\item{pred_S_e}{(Optional) Predicted \eqn{P(S_e = 1 \mid R)}, \eqn{\text{PRED}_{S_e}(R)}.}

\item{pred_S_o}{(Optional) Predicted \eqn{P(S_o = 1 \mid R)}, \eqn{\text{PRED}_{S_o}(R)}.}

\item{theta_init}{Initial estimate of the treatment effect on the train data.}

\item{eps}{Small constant for numerical stability of \code{sigma2} estimate (default \code{1e-2}).}

\item{ml_params}{List of parameters for random forest:
\describe{
  \item{ntree}{Number of trees}
  \item{classwt_Y}{Class weights for \code{pred_Y} model; default \code{c(10, 1)})}
  \item{seed}{User specified seed passed to each \code{ranger} function for reproducibility; default \code{NULL}}
  \item{cores}{Number of cores used by `ranger` for parallel training; default \code{1}}
  \item{nfolds}{Number of folds for cross-fitting (default \code{5}).}
  \item{train_ratio}{Proportion for training in sample split (default \code{0.5}).}
}}

\item{se}{Logical; compute standard errors via bootstrap? (default \code{TRUE}).}

\item{se_params}{List of bootstrap parameters:
\describe{
  \item{B}{Number of bootstrap replications (default \code{1000})}
  \item{fix_seed}{If \code{TRUE}, deterministic seeding is used with `set.seed(b)` for the *b*‑th replication (default \code{FALSE})}
  \item{cores}{Number of cores for bootstrap replications (default \code{1}).}
  \item{clusters}{Clusters for the bootstrap. If \code{NULL}, uses individual-level bootstrap}
}}
}
\description{
Uses all data as a test set
}
\keyword{internal}
