% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.cv.rsv}
\alias{summary.cv.rsv}
\title{Summary method for cv.rsv objects}
\usage{
\method{summary}{cv.rsv}(object, ...)
}
\arguments{
\item{object}{An object of class "cv.rsv"}

\item{...}{Additional arguments (unused)}
}
\description{
Summary method for cv.rsv objects
}
