% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.rsv}
\alias{summary.rsv}
\title{Summary method for rsv objects}
\usage{
\method{summary}{rsv}(object, ...)
}
\arguments{
\item{object}{An object of class "rsv"}

\item{...}{Additional arguments (unused)}
}
\description{
Summary method for rsv objects
}
