% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_bootstrap_params}
\alias{validate_bootstrap_params}
\title{Validate bootstrap parameters}
\usage{
validate_bootstrap_params(B, clusters, n)
}
\arguments{
\item{B}{Number of bootstrap replications}

\item{clusters}{Cluster identifiers}

\item{n}{Sample size}
}
\description{
Validate bootstrap parameters
}
\keyword{internal}
