% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_cv_params}
\alias{validate_cv_params}
\title{Validate cross-fitting parameters}
\usage{
validate_cv_params(nfolds, n)
}
\arguments{
\item{nfolds}{Number of folds}

\item{n}{Sample size}
}
\description{
Validate cross-fitting parameters
}
\keyword{internal}
