% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_models}
\alias{validate_models}
\title{Validate model specifications}
\usage{
validate_models(models)
}
\arguments{
\item{models}{List of model specifications}
}
\description{
Validate model specifications
}
\keyword{internal}
