% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_predictions}
\alias{validate_predictions}
\title{Validate predictions}
\usage{
validate_predictions(pred_Y, pred_D, pred_S_e, pred_S_o, n)
}
\arguments{
\item{pred_Y}{Predicted Y}

\item{pred_D}{Predicted D}

\item{pred_S_e}{Predicted S_e}

\item{pred_S_o}{Predicted S_o}

\item{n}{Expected length}
}
\description{
Validate predictions
}
\keyword{internal}
