% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_R}
\alias{validate_R}
\title{Validate remote sensing variables}
\usage{
validate_R(R, n)
}
\arguments{
\item{R}{Remote sensing variables (matrix or data frame)}

\item{n}{Expected number of observations}
}
\description{
Validate remote sensing variables
}
\keyword{internal}
