% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_required_inputs}
\alias{validate_required_inputs}
\title{Validate required inputs}
\usage{
validate_required_inputs(Y, D, S_e, S_o)
}
\arguments{
\item{Y}{Outcome vector}

\item{D}{Treatment vector}

\item{S_e}{Experimental sample indicator}

\item{S_o}{Observational sample indicator}
}
\description{
Validate required inputs
}
\keyword{internal}
