% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_split_params}
\alias{validate_split_params}
\title{Validate sample split parameters}
\usage{
validate_split_params(train.proportion)
}
\arguments{
\item{train.proportion}{Training proportion}
}
\description{
Validate sample split parameters
}
\keyword{internal}
