% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_X}
\alias{validate_X}
\title{Validate discrete covariates}
\usage{
validate_X(X, n, X_cols = NULL)
}
\arguments{
\item{X}{Discrete covariates (matrix or data frame)}

\item{n}{Expected number of observations}

\item{X_cols}{Column names for X (optional)}
}
\description{
Validate discrete covariates
}
\keyword{internal}
