% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{vcov.cv.rsv}
\alias{vcov.cv.rsv}
\title{Extract variance-covariance matrix from cv.rsv objects}
\usage{
\method{vcov}{cv.rsv}(object, ...)
}
\arguments{
\item{object}{An object of class "cv.rsv"}

\item{...}{Additional arguments (unused)}
}
\value{
1x1 variance-covariance matrix
}
\description{
Extract variance-covariance matrix from cv.rsv objects
}
