% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{vcov.rsv}
\alias{vcov.rsv}
\title{Extract variance-covariance matrix from rsv objects}
\usage{
\method{vcov}{rsv}(object, ...)
}
\arguments{
\item{object}{An object of class "rsv"}

\item{...}{Additional arguments (unused)}
}
\value{
1x1 variance-covariance matrix
}
\description{
Extract variance-covariance matrix from rsv objects
}
