% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsv_methods.R
\name{print.rsv}
\alias{print.rsv}
\title{Print method for rsv objects}
\usage{
\method{print}{rsv}(x, ...)
}
\arguments{
\item{x}{An object of class "rsv".}

\item{...}{Additional arguments (unused).}
}
\description{
Print method for rsv objects
}
